asset.get_bank_visual_identity/requirements.txt
--extra-index-url https://backend:dc222fa6-452f-4253-a2fd-a52e7ba2e7e9@nexus.sigame.com.br/repository/pypi/simple
etria-logger==0.1.8
mnemosine==0.1.2
boto3==1.21.21
flask==2.0.3
pydantic==1.9.0
python-decouple==3.6
hypercorn==0.13.2

uvloop==0.17.0
uvicorn==0.20.0
asgiref==3.5.2
hypercorn==0.13.2